refactor: 패키지 이름 수정
- 배치가 추가됨에 따라 api, batch, domain, global로 패키지 구분
- 모든 파일에 해당 내용 적용
