+++ src/main/java/com/evinfo/api/charger/dto/client/ChargerClientResponseDto.java
@@ -1,8 +1,8 @@
-package com.evinfo.domain.charger.dto.client;
+package com.evinfo.api.charger.dto.client;
 
-import com.evinfo.domain.charger.domain.Charger;
-import com.evinfo.domain.charger.domain.ChargerStat;
-import com.evinfo.domain.charger.domain.ChargerType;
+import com.evinfo.domain.charger.Charger;
+import com.evinfo.domain.charger.ChargerStat;
+import com.evinfo.domain.charger.ChargerType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
