+++ src/main/java/com/evinfo/api/charger/repository/ChargerRepository.java
@@ -1,6 +1,6 @@
-package com.evinfo.domain.charger.repository;
+package com.evinfo.api.charger.repository;
 
-import com.evinfo.domain.charger.domain.Charger;
+import com.evinfo.domain.charger.Charger;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChargerRepository extends JpaRepository<Charger, Long> {
