+++ src/test/java/com/evinfo/api/charger/controller/ChargerControllerTest.java
@@ -1,9 +1,9 @@
-package com.evinfo.domain.charger.controller;
+package com.evinfo.api.charger.controller;
 
+import com.evinfo.api.charger.dto.ChargerResponseDto;
+import com.evinfo.api.charger.service.ChargerClient;
+import com.evinfo.api.charger.service.ChargerService;
 import com.evinfo.docs.Documentation;
-import com.evinfo.domain.charger.dto.ChargerResponseDto;
-import com.evinfo.domain.charger.service.ChargerClient;
-import com.evinfo.domain.charger.service.ChargerService;
 import com.evinfo.utils.ChargerGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
