+++ src/test/java/com/evinfo/api/charger/service/ChargerServiceTest.java
@@ -1,9 +1,9 @@
-package com.evinfo.domain.charger.service;
+package com.evinfo.api.charger.service;
 
-import com.evinfo.domain.charger.domain.Charger;
-import com.evinfo.domain.charger.dto.ChargerRequestDto;
-import com.evinfo.domain.charger.dto.ChargerResponseDto;
-import com.evinfo.domain.charger.repository.ChargerRepository;
+import com.evinfo.api.charger.dto.ChargerRequestDto;
+import com.evinfo.api.charger.dto.ChargerResponseDto;
+import com.evinfo.api.charger.repository.ChargerRepository;
+import com.evinfo.domain.charger.Charger;
 import com.evinfo.utils.ChargerGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
